Use 0.3-beta.1 build
